requirements.txt
Flask==3.0.3
Flask-Babel==4.0.0
Babel==2.15.0
Pillow==10.4.0
gunicorn==21.2.0